added bulk email, fixed calendar

+++ FIT5032_A/Controllers/CoursController.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using FIT5032_A.Models;
+using Microsoft.AspNet.Identity;
 
 namespace FIT5032_A.Controllers
 {
@@ -17,8 +18,23 @@ namespace FIT5032_A.Controllers
         // GET: Cours
         public ActionResult Index()
         {
-            var courses = db.Courses.Include(c => c.Language).Include(c => c.School);
-            return View(courses.ToList());
+            List<Cours> courses = new List<Cours>();
+            if (User.IsInRole("Student"))
+            {
+                var userId = User.Identity.GetUserId();
+                Student student = db.Students.Where(s => s.UserId == userId).First();
+                var studentenrolments = db.Enrolments.Where(e => e.StudentId == student.Id).ToList();
+                foreach (Enrolment enrolment in studentenrolments)
+                {
+                    Cours course = db.Courses.Where(c => c.Id == enrolment.CourseId).First();
+                    courses.Add(course);
+                }
+            }
+            else
+            {
+                courses = db.Courses.Include(c => c.Language).Include(c => c.School).ToList();
+            }
+            return View(courses);
         }
 
         // GET: Cours/Details/5
@@ -37,10 +53,14 @@ namespace FIT5032_A.Controllers
         }
 
         // GET: Cours/Create
+        [Authorize]
         public ActionResult Create()
         {
-            ViewBag.LanguageId = new SelectList(db.Languages, "Id", "Name");
-            ViewBag.SchoolId = new SelectList(db.Schools, "Id", "Name");
+            if (User.IsInRole("Administrator"))
+            {
+                ViewBag.LanguageId = new SelectList(db.Languages, "Id", "Name");
+                ViewBag.SchoolId = new SelectList(db.Schools, "Id", "Name");
+            }
             return View();
         }
 
@@ -49,13 +69,18 @@ namespace FIT5032_A.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public ActionResult Create([Bind(Include = "Id,Name,Description,Start,End,LanguageId,SchoolId")] Cours cours)
         {
             if (ModelState.IsValid)
             {
-                db.Courses.Add(cours);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (User.IsInRole("Administrator"))
+                {
+                    db.Courses.Add(cours);
+                    db.SaveChanges();
+                   
+                    return RedirectToAction("Index");
+                }
             }
 
             ViewBag.LanguageId = new SelectList(db.Languages, "Id", "Name", cours.LanguageId);
@@ -64,19 +89,25 @@ namespace FIT5032_A.Controllers
         }
 
         // GET: Cours/Edit/5
+        [Authorize]
         public ActionResult Edit(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Cours cours = db.Courses.Find(id);
-            if (cours == null)
+            Cours cours = new Cours();
+
+            if (User.IsInRole("Administrator"))
             {
-                return HttpNotFound();
-            }
-            ViewBag.LanguageId = new SelectList(db.Languages, "Id", "Name", cours.LanguageId);
-            ViewBag.SchoolId = new SelectList(db.Schools, "Id", "Name", cours.SchoolId);
+                cours = db.Courses.Find(id);
+                if (cours == null)
+                {
+                    return HttpNotFound();
+                }
+                ViewBag.LanguageId = new SelectList(db.Languages, "Id", "Name", cours.LanguageId);
+                ViewBag.SchoolId = new SelectList(db.Schools, "Id", "Name", cours.SchoolId);
+                    }
             return View(cours);
         }
 
@@ -85,30 +116,39 @@ namespace FIT5032_A.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public ActionResult Edit([Bind(Include = "Id,Name,Description,Start,End,LanguageId,SchoolId")] Cours cours)
         {
-            if (ModelState.IsValid)
+            if (User.IsInRole("Administrator"))
             {
-                db.Entry(cours).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(cours).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                ViewBag.LanguageId = new SelectList(db.Languages, "Id", "Name", cours.LanguageId);
+                ViewBag.SchoolId = new SelectList(db.Schools, "Id", "Name", cours.SchoolId);
             }
-            ViewBag.LanguageId = new SelectList(db.Languages, "Id", "Name", cours.LanguageId);
-            ViewBag.SchoolId = new SelectList(db.Schools, "Id", "Name", cours.SchoolId);
             return View(cours);
         }
 
         // GET: Cours/Delete/5
+        [Authorize]
         public ActionResult Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Cours cours = db.Courses.Find(id);
-            if (cours == null)
+            Cours cours = new Cours();
+            if (User.IsInRole("Administrator"))
             {
-                return HttpNotFound();
+                cours = db.Courses.Find(id);
+                if (cours == null)
+                {
+                    return HttpNotFound();
+                }
             }
             return View(cours);
         }
@@ -116,11 +156,15 @@ namespace FIT5032_A.Controllers
         // POST: Cours/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public ActionResult DeleteConfirmed(int id)
         {
-            Cours cours = db.Courses.Find(id);
-            db.Courses.Remove(cours);
-            db.SaveChanges();
+            if (User.IsInRole("Administrator"))
+            {
+                Cours cours = db.Courses.Find(id);
+                db.Courses.Remove(cours);
+                db.SaveChanges();
+            }
             return RedirectToAction("Index");
         }
